requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.1
alembic==1.12.1
pytest==7.4.3
asyncpg==0.29.0
sqlalchemy-utils==0.41.1
psycopg2-pool==1.1 